+++ CHANGELOG.md
@@ -3,7 +3,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
-### 🔄 Changed
+### ✅ Added
+- Factory method for customizing the search results view
 
 # [4.47.1](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.47.1)
 _January 29, 2024_